Time: 0 ms (100.00%), Space: 5.9 MB (71.51%) - LeetHub

+++ 62-unique-paths/62-unique-paths.cpp
@@ -5,7 +5,7 @@ public:
         int N = n + m - 2; 
         int r = min(m-1,n-1);
         
-        double res =1;
+        long res =1;
         for(int i=1;i<=r;i++){
             res=res*(N-r+i)/i;
         }
